tone_analyzer/requirements.txt
requests
wrapt
python-keystoneclient
python-swiftclient
Flask
